requirements.txt
pillow
python-telegram-bot
apscheduler
requests